AM-2894 refactor test ORM refresh endpoint
   [AM-2894](https://tools.hmcts.net/jira/browse/AM-2894)

Refactor the test ORM refresh endpoint:
* rename
* enable swagger
* hide/disable behind `TESTING_SUPPORT_ENABLED` env variable.
