+++ charts/am-org-role-mapping-service/values.aat.template.yaml
@@ -6,3 +6,4 @@ java:
     ENABLE_CACHE: none
     ORG_ROLE_MAPPING_DB_HOST: am-org-role-mapping-service-postgres-db-v11-{{ .Values.global.environment }}.postgres.database.azure.com
     ORG_ROLE_MAPPING_DB_USERNAME: am@am-org-role-mapping-service-postgres-db-v11-{{ .Values.global.environment }}
+    TESTING_SUPPORT_ENABLED: true
