+++ charts/am-org-role-mapping-service/values.preview.template.yaml
@@ -49,6 +49,7 @@ java:
     ORG_ROLE_MAPPING_DB_PASSWORD: "{{ .Values.postgresql.auth.password}}"
     ORG_ROLE_MAPPING_DB_PORT: "{{ .Values.postgresql.auth.port}}"
     ORG_ROLE_MAPPING_DB_OPTIONS: "?stringtype=unspecified"
+    TESTING_SUPPORT_ENABLED: true
     RUN_LD_ON_STARTUP: true
   postgresql:
     enabled: true
