+++ src/integrationTest/java/uk/gov/hmcts/reform/orgrolemapping/config/SwaggerPublisher.java
@@ -2,10 +2,12 @@ package uk.gov.hmcts.reform.orgrolemapping.config;
 
 import org.junit.Before;
 import org.junit.Test;
+import org.springframework.test.context.TestPropertySource;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.setup.MockMvcBuilders;
 import org.springframework.web.context.WebApplicationContext;
 import uk.gov.hmcts.reform.orgrolemapping.controller.BaseTest;
+
 import javax.inject.Inject;
 import java.io.OutputStream;
 import java.nio.file.Files;
@@ -20,6 +22,10 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
  * Each run of workflow .github/workflows/swagger.yml on master should automatically save and upload (if updated)
  * documentation.
  */
+@TestPropertySource(properties = {
+    // NB: hide testing-support endpoint from Swagger Publish
+    "testing.support.enabled=false"
+})
 public class SwaggerPublisher extends BaseTest {
 
     private MockMvc mockMvc;
