+++ src/main/java/uk/gov/hmcts/reform/orgrolemapping/servicebus/JRDMessagingConfiguration.java
@@ -6,7 +6,7 @@ import com.azure.messaging.servicebus.ServiceBusSenderClient;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang.StringUtils;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnExpression;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Primary;
@@ -32,7 +32,7 @@ public class JRDMessagingConfiguration {
     String environment;
 
     @Bean("jrdPublisher")
-    @ConditionalOnProperty(name = "amqp.jrd.enabled", havingValue = "true")
+    @ConditionalOnExpression("${testing.support.enabled} && ${amqp.jrd.enabled}")
     public ServiceBusSenderClient getServiceBusSenderClient() {
         log.debug("Getting the ServiceBusSenderClient in JRD");
         logServiceBusVariables();
